Tests/LibWeb/Text/expected/IntersectionObserver/observe-box-inside-container-with-scrollable-overflow.txt

The box is not visible.
The box is now visible.
